Fix VGLLine() in depths > 8.
It started truncating its color arg to 8 bits using plot() in r229415.
The version in r229415 is also more than 3 times slower in segmented
modes, by doing more syscalls to move the window.

+++ lib/libvgl/simple.c
@@ -248,7 +248,7 @@ VGLGetXY(VGLBitmap *object, int x, int y)
 #define SL_ABSOLUTE(i,j,k)     ( (i-j)*(k = ( (i-j)<0 ? -1 : 1)))
 
 void
-plot(VGLBitmap * object, int x, int y, int flag, byte color)
+plot(VGLBitmap * object, int x, int y, int flag, u_long color)
 {
   /* non-zero flag indicates the pixels need swapping back. */
   if (flag)
